Crate service_modules

Structs

AddBlueprint

Blueprint

Hash

Functions

blueprint_file_name

Calculates filename of the blueprint

blueprint_fname

extract_module_file_name

Return file name with .wasm extension stripped. None if extension wasn’t .wasm

is_blueprint

Returns true if file is named like a blueprint would be

is_module_wasm

is_service

load_module

module_config

module_config_name_hash

Calculates filename of the config for a wasm module

module_config_name_json

Calculates filename of the config for a wasm module, given a hash or name of the module.

module_file_name

Calculates the name of a wasm module file, given a hash or name of the module.

module_file_name_hash

Calculates the name of a wasm module file, given a hash of the module.

service_file_name